Fix warnings up to -Wsign-conversion

+++ evaluation/performance/gpuConv.cpp
@@ -13,17 +13,18 @@ struct EntryConvCPU : public PerformanceTaskDefault {
     PerformanceMeasure& pm = this->measure.back();
     if (verbose) std::cout << "    Running " << this->name << " ... ";
     cv::Mat tmp;
-    uint64 start = 0;
+    int64 start = 0;
     cv::GaussianBlur(src, tmp, cv::Size(KS, KS), 0);
     for (int i = 0; i != runs; ++i) {
       start = cv::getTickCount();
       cv::GaussianBlur(src, tmp, cv::Size(KS, KS), 0);
-      pm.measures.push_back(cv::getTickCount() - start);
+      pm.measures.push_back(static_cast<uint64>(cv::getTickCount() - start));
     }
     if (verbose)
       std::cout << std::setprecision(3)
-                << static_cast<double>((cv::getTickCount() - start) * 1000) / (runs * cv::getTickFrequency()) << "ms"
-                << std::endl;
+                << static_cast<double>((cv::getTickCount() - start) * 1000) /
+                       (runs * static_cast<double>(cv::getTickFrequency()))
+                << "ms" << std::endl;
   }
 };
 
@@ -36,7 +37,7 @@ struct EntryConvCL : public PerformanceTaskDefault {
     this->measure.push_back(PerformanceMeasure(src_name, this->name, src.cols, src.rows));
     PerformanceMeasure& pm = this->measure.back();
     if (verbose) std::cout << "    Running " << this->name << " ... ";
-    uint64 start = 0;
+    int64 start = 0;
     cv::Mat tmp2;
     cv::UMat in = src.getUMat(cv::ACCESS_READ), tmp;  // to GPU
     cv::GaussianBlur(in, tmp, cv::Size(KS, KS), 0);
@@ -49,12 +50,13 @@ struct EntryConvCL : public PerformanceTaskDefault {
       cv::GaussianBlur(in, tmp, cv::Size(KS, KS), 0);
       tmp.copyTo(tmp2);  // to RAM
 
-      pm.measures.push_back(cv::getTickCount() - start);
+      pm.measures.push_back(static_cast<uint64>(cv::getTickCount() - start));
     }
     if (verbose)
       std::cout << std::setprecision(3)
-                << static_cast<double>((cv::getTickCount() - start) * 1000) / (runs * cv::getTickFrequency()) << "ms"
-                << std::endl;
+                << static_cast<double>((cv::getTickCount() - start) * 1000) /
+                       (runs * static_cast<double>(cv::getTickFrequency()))
+                << "ms" << std::endl;
   }
 };
 
@@ -70,19 +72,20 @@ struct EntryConvCLNT : public PerformanceTaskDefault {
     cv::Mat tmp2;
     cv::GaussianBlur(in, tmp, cv::Size(KS, KS), 0);
     tmp.copyTo(tmp2);  // to RAM
-    uint64 start = 0;
+    int64 start = 0;
     for (int i = 0; i != runs; ++i) {
       in = src.getUMat(cv::ACCESS_READ);
       start = cv::getTickCount();
       cv::GaussianBlur(in, tmp, cv::Size(KS, KS), 0);
 
-      pm.measures.push_back(cv::getTickCount() - start);
+      pm.measures.push_back(static_cast<uint64>(cv::getTickCount() - start));
       tmp.copyTo(tmp2);  // to RAM
     }
     if (verbose)
       std::cout << std::setprecision(3)
-                << static_cast<double>((cv::getTickCount() - start) * 1000) / (runs * cv::getTickFrequency()) << "ms"
-                << std::endl;
+                << static_cast<double>((cv::getTickCount() - start) * 1000) /
+                       (runs * static_cast<double>(cv::getTickFrequency()))
+                << "ms" << std::endl;
   }
 };
 
@@ -97,7 +100,7 @@ struct EntryConvCuda : public PerformanceTaskDefault {
     PerformanceMeasure& pm = this->measure.back();
     if (verbose) std::cout << "    Running " << this->name << " ... ";
     cv::Mat tmp2;
-    uint64 start = 0;
+    int64 start = 0;
     cv::Ptr<cv::cuda::Filter> gauss = cv::cuda::createGaussianFilter(src.type(), src.type(), cv::Size(KS, KS), 0);
     cv::cuda::GpuMat in, tmp;
     in.upload(src);
@@ -112,8 +115,9 @@ struct EntryConvCuda : public PerformanceTaskDefault {
     }
     if (verbose)
       std::cout << std::setprecision(3)
-                << static_cast<double>((cv::getTickCount() - start) * 1000) / (runs * cv::getTickFrequency()) << "ms"
-                << std::endl;
+                << static_cast<double>((cv::getTickCount() - start) * 1000) /
+                       (runs * static_cast<double>(cv::getTickFrequency()))
+                << "ms" << std::endl;
   }
 };
 
@@ -125,7 +129,7 @@ struct EntryConvCudaNT : public PerformanceTaskDefault {
     this->measure.push_back(PerformanceMeasure(src_name, this->name, src.cols, src.rows));
     PerformanceMeasure& pm = this->measure.back();
     if (verbose) std::cout << "    Running " << this->name << " ... ";
-    uint64 start = 0;
+    int64 start = 0;
     cv::Ptr<cv::cuda::Filter> gauss = cv::cuda::createGaussianFilter(src.type(), src.type(), cv::Size(KS, KS), 0);
     cv::cuda::GpuMat in, tmp;
     in.upload(src);
@@ -137,8 +141,9 @@ struct EntryConvCudaNT : public PerformanceTaskDefault {
     }
     if (verbose)
       std::cout << std::setprecision(3)
-                << static_cast<double>((cv::getTickCount() - start) * 1000) / (runs * cv::getTickFrequency()) << "ms"
-                << std::endl;
+                << static_cast<double>((cv::getTickCount() - start) * 1000) /
+                       (runs * static_cast<double>(cv::getTickFrequency()))
+                << "ms" << std::endl;
   }
 };
 #endif
